services/local-inference/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.35,<1.0
torch>=2.4
diffusers>=0.35
transformers>=4.55
accelerate>=1.10
Pillow>=11.0
imageio>=2.37
imageio-ffmpeg>=0.6
numpy>=2.2
safetensors>=0.5
sentencepiece>=0.2
protobuf>=5.0
